requirements.txt
streamlit
openai
langchain
pdfminer.six
python-dotenv
PyMuPDF